Move utils to test file, inlined methods and contract name
